requirements.txt
apparmor==3.0.4
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
Brlapi==0.8.3
bs4==0.0.2
capstone==6.0.0a1
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.1.7
colorama==0.4.6
colored-traceback==0.4.2
command-not-found==0.3
crypto==1.4.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
fail2ban==0.11.2
Flask==3.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
intervaltree==3.1.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
LibAppArmor==3.0.4
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
more-itertools==8.10.0
mysql-connector-python==9.2.0
Naked==0.1.32
netifaces==0.11.0
networkx==2.4
numpy==1.21.5
oauthlib==3.2.0
packaging==24.2
paramiko==3.5.0
pillow==11.2.1
plumbum==1.9.0
plyer==2.1.0
protobuf==3.12.4
psutil==6.1.0
pwntools==4.13.1
pyasn1==0.6.1
pycairo==1.20.1
pycparser==2.22
pycryptodome==3.21.0
pycups==2.0.1
pyelftools==0.31
Pygments==2.18.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==4.3.1
requests==2.25.1
ROPGadget==7.5
rpyc==6.0.1
rsa==4.9
SecretStorage==3.3.1
selinux==3.3
sepolicy==3.3
setools==4.4.0
shellescape==3.8.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.6
ssh-import-id==5.11
systemd-python==234
ufw==0.36.1
unicorn==2.1.1
unix-ar==0.2.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.1.3
xdg==5
zipp==1.0.0
zstandard==0.23.0
